Add files via upload
This code opens the file "Rosebowl.txt" for reading and appending using a with statement. 
It first adds the winners from 2014 to 2022 to the end of the file by creating a list of winners and 
using the join method to concatenate them with newlines. It then moves the file pointer to the beginning 
of the file and loops through each line, counting the number of wins for each team using a dictionary. 
It filters the dictionary to include only teams with 4 or more wins and **Creative element** alphabetizes the resulting dictionary 
by team name. Finally, it prints the updated list of winners with the number of wins for each team.
The try block is used to catch any IOErrors that may occur when attempting to open the file. If an IOError 
occurs, the code prints an error message.